Move computeMLDSAMu helper to TestUtil
