--- tst/com/amazon/corretto/crypto/provider/test/PublicUtilsTest.java
@@ -1,56 +1,0 @@
-// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-package com.amazon.corretto.crypto.provider.test;
-
-import static org.junit.jupiter.api.Assertions.assertArrayEquals;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import com.amazon.corretto.crypto.provider.AmazonCorrettoCryptoProvider;
-import com.amazon.corretto.crypto.provider.PublicUtils;
-import java.security.KeyFactory;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.Provider;
-import java.security.PublicKey;
-import java.security.spec.X509EncodedKeySpec;
-import java.util.Arrays;
-import org.junit.jupiter.api.condition.DisabledIf;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.api.parallel.Execution;
-import org.junit.jupiter.api.parallel.ExecutionMode;
-import org.junit.jupiter.api.parallel.ResourceAccessMode;
-import org.junit.jupiter.api.parallel.ResourceLock;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-
-@Execution(ExecutionMode.CONCURRENT)
-@ExtendWith(TestResultLogger.class)
-@ResourceLock(value = TestUtil.RESOURCE_GLOBAL, mode = ResourceAccessMode.READ)
-public class PublicUtilsTest {
-  private static final Provider NATIVE_PROVIDER = AmazonCorrettoCryptoProvider.INSTANCE;
-
-  // TODO: remove this disablement when ACCP consumes an AWS-LC-FIPS release with ML-DSA
-  private static boolean mlDsaDisabled() {
-    return AmazonCorrettoCryptoProvider.INSTANCE.isFips()
-        && !AmazonCorrettoCryptoProvider.INSTANCE.isExperimentalFips();
-  }
-
-  @ParameterizedTest
-  @ValueSource(strings = {"ML-DSA-44", "ML-DSA-65", "ML-DSA-87"})
-  @DisabledIf("mlDsaDisabled")
-  public void testComputeMLDSAExtMu(String algorithm) throws Exception {
-    KeyPair keyPair = KeyPairGenerator.getInstance(algorithm, NATIVE_PROVIDER).generateKeyPair();
-    PublicKey nativePub = keyPair.getPublic();
-    KeyFactory bcKf = KeyFactory.getInstance("ML-DSA", TestUtil.BC_PROVIDER);
-    PublicKey bcPub = bcKf.generatePublic(new X509EncodedKeySpec(nativePub.getEncoded()));
-
-    byte[] message = new byte[256];
-    Arrays.fill(message, (byte) 0x41);
-    byte[] mu = PublicUtils.computeMLDSAMu(nativePub, message);
-    assertEquals(64, mu.length);
-    // We don't have any other implementations of mu calculation to test against, so just assert
-    // that mu is equivalent
-    // generated from both ACCP and BouncyCastle keys.
-    assertArrayEquals(mu, PublicUtils.computeMLDSAMu(bcPub, message));
-  }
-}
